\name{benchmarks}
\alias{benchmark_1}
\alias{benchmark_1_0.05}
\alias{benchmark_2}
\alias{benchmark_3}

\docType{data}
\title{
Summary results from some benchmarks reported in the vignette.
}

\description{

  Summary results from some benchmarks reported in the
  vignette. Included are timings, sizes of return objects and key
  oputput from each simulation.

  They are here mainly to facilitate creation of table from the vignette
  itself. The scripts are available under "inst/miscell".

}
\usage{
data(benchmark_1)
data(benchmark_1_0.05)
data(benchmark_2)
data(benchmark_3)

}
\format{
  Data frames.
}


\examples{
data(benchmark_1)
benchmark_1
}
\keyword{datasets}


