\name{OncoSimulR-package}
\alias{OncoSimulR-package}
\alias{OncoSimulR}
\docType{package}
\title{
Simulation of cancer progresion with order restrictions
}
\description{
Functions for simulating cancer progression data, including drivers and passengers, and allowing for order restrictions. 
}
\details{
\tabular{ll}{
Package: \tab OncoSimulR\cr
Type: \tab Package\cr
Version: \tab 0.99\cr
Date: \tab 2014-05-22\cr
License: \tab GPL (>= 3)\cr
Depends: \tab R (>= 3.1.0)\cr
Imports: \tab Rcpp (>= 0.11.1), parallel, data.table, graph\cr
LinkingTo: \tab Rcpp\cr
## SystemRequirements: \tab C++11\cr
Built: \tab R 3.1.0; x86_64-pc-linux-gnu; 2014-06-10 17:06:25.914 UTC; unix\cr
}

Index:
\preformatted{
examplePosets           Example posets
oncoSimulIndiv          Simulate tumor progression for one or more
                        individuals.
plot.oncosimul          Plot simulated tumor progression data.
plotPoset               Plot a poset.
poset                   Poset
samplePop               Obtain a sample from a population of
                        simulations.
}
}
\author{
Ramon Diaz-Uriarte. 

Maintainer: Ramon Diaz-Uriarte <ramon.diaz@iib.uam.es>
}
\keyword{ package }






